+++ internal/indexedimage/kmeans.go
@@ -1,4 +1,4 @@
-package main
+package indexedimage
 
 import (
 	"fmt"
