+++ internal/indexedimage/quantizer.go
@@ -1,4 +1,4 @@
-package main
+package indexedimage
 
 import (
 	"math"
